+++ emuDQM/emuMonitor/include/i2oEmuMonitorMsg.h
@@ -8,15 +8,15 @@
 
 typedef struct _I2O_EMUMONITOR_CREDIT_MESSAGE_FRAME {
   I2O_PRIVATE_MESSAGE_FRAME PvtMessageFrame;
-  unsigned long nEventCredits;
-  unsigned long prescalingFactor;
+  uint32_t nEventCredits;
+  uint32_t prescalingFactor;
 } I2O_EMUMONITOR_CREDIT_MESSAGE_FRAME, *PI2O_EMUMONITOR_CREDIT_MESSAGE_FRAME;
 
 typedef struct _I2O_EMU_DATA_MESSAGE_FRAME {
   I2O_PRIVATE_MESSAGE_FRAME PvtMessageFrame;
-  unsigned long  runNumber;
-  unsigned long  nEventCreditsHeld;
-  unsigned short errorFlag; // for DQM; filled by EmuRUI's server only
+  uint32_t  runNumber;
+  uint32_t  nEventCreditsHeld;
+  uint32_t errorFlag; // for DQM; filled by EmuRUI's server only
 } I2O_EMU_DATA_MESSAGE_FRAME, *PI2O_EMU_DATA_MESSAGE_FRAME;
 
 #endif
